:recycle: Split enormous test file
